feat: Add HTMX demo for OID4VC
After talking with @dbluhm, we've decided to replace the
heavy/time-consuming react.js demo with a much lighter weight and faster
HTMX-based demo. What's contained within this commit is a first pass
towards those efforts and (for the most part) works just fine. Future
commits will expound upon the presentation flow, as well as clean up the
code to make it a bit easier to read.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>