feat: Upgrade oid4vc to draft 13 & mdl test progress
In an effort to push the mDL integration tests as far as possible, I
have made the tweaks necessary to get the Sphereon test harness to
recognize the service as adhering to the oid4vc draft 13 spec. I have
also pushed the credo tests as far as I possibly could to the point
where we are now seeing the same error that we're seeing within the
Sphereon mobile app. Despite these changes, the Paradym wallet is still
refusing to recognize the service as a valid oid4vc service.

Interop Integration test status:
- Sphereon now passes
- Credo now complains about receiving all the data for the credential
  but not able to read everything needed to decode. If I understood the
  error correctly that is. It's late and I'm tired.

Great progress made today!

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>